Update Device_Type_8_Zone_Irrigation_Controller.groovy
Adds the "Momentary" capability with the associated push command.  The push command sends a message to arduino to "advance" which advances to the next station.

This allows a SmartApp to advance the relays to the next Zone.  For example, you could connect a remote device that could send a switch.on command to turn on all the stations and then a momentary.push command to advance the stations through the complete allOn sequence.  I am using this with a custom Aeon Minimote smart app to let a irrigation repair man blow out my sprinklers using the mini mote to control the system while I am at the office.