Update Device_Type_24_Zone_Irrigation_Controller.groovy
Adds momentary capability and the associated push() command.  This capability is associated with the Arduino to advance the relays one by one through the remaining queue until the queue is finished.  There is no associated device tile (not needed since you can each zone tile to deactivate.  Its provided so it can be used by SmartApps