Update Device_Type_24_Zone_Irrigation_Controller.groovy
The 24 zone irrigation scheduler presented an interesting challenge.  We use one concatenated message to send all the station zone times from the device type to the Arduino.  This was required because if you try to send back to back messages,  for example trying to send 24 messages, the earlier messages are dropped as subsequent messages are sent over.  However, with 24 zones, our concatenated message exceeded an apparent 63 character payload limit.  This means we needed to send two messages.  A special call out to SmartThings support for helping identify a method to send multiple Zigbee messages by introducing a delay between the messages.  The delay is coded into the actual command that is sent to the Zigbee chip, which parses the command and executes the delay.  This is a critical bug fix.