Update IrrigationSchedulerApp.groovy
This is a bug fix.  In cases where you do not check the weather (you have turned off the sliders in the smart app), the previous code was throwing an error (trying to perform a rounding function on an integer).   If you do not plan to use the weather api, this is a critical bug fix for your smart app to work.  Advanced the version to 3.05